Check if any conflicting_daemon_pod exists

+++ deploy-disconnected-registry/deploy.sh
@@ -50,28 +50,33 @@ function side_evict_error() {
         echo "No masters on ${status}"
     else
         conflicting_daemon_pod=$(oc --kubeconfig=${KUBEC} get pod -n openshift-machine-config-operator -o wide --no-headers | grep daemon | grep ${conflicting_node} | cut -f1 -d\ )
-        pattern_1="$(oc --kubeconfig=${KUBEC} logs -n openshift-machine-config-operator ${conflicting_daemon_pod} -c machine-config-daemon | grep drain.go | grep evicting | tail -1 | grep pods)"
-        pattern_2="$(oc --kubeconfig=${KUBEC} logs -n openshift-machine-config-operator ${conflicting_daemon_pod} -c machine-config-daemon | grep drain.go | grep "Draining failed" | tail -1 | grep pod)"
 
-        for log_entry in "${pattern_1}" "${pattern_2}"; do
-            if [[ -z ${log_entry} ]]; then
-                echo "No Conflicting LogEntry on ${conflicting_daemon_pod}"
-            else
-                echo ">> Conflicting LogEntry Found!!"
-                pod=$(echo ${log_entry##*pods/} | cut -d\" -f2)
-                conflicting_ns=$(oc --kubeconfig=${KUBEC} get pod -A | grep ${pod} | cut -f1 -d\ )
+        # Check if conflicting_daemon_pod is not empty
+        if [[ -z ${conflicting_daemon_pod} ]]; then
+            echo "No conflicting daemon pod exists in ${conflicting_node}"
+        else
+            pattern_1="$(oc --kubeconfig=${KUBEC} logs -n openshift-machine-config-operator ${conflicting_daemon_pod} -c machine-config-daemon | grep drain.go | grep evicting | tail -1 | grep pods)"
+            pattern_2="$(oc --kubeconfig=${KUBEC} logs -n openshift-machine-config-operator ${conflicting_daemon_pod} -c machine-config-daemon | grep drain.go | grep "Draining failed" | tail -1 | grep pod)"
 
-                echo ">> Clean Eviction triggered info: "
-                echo NODE: ${conflicting_node}
-                echo DAEMON: ${conflicting_daemon_pod}
-                echo NS: ${conflicting_ns}
-                echo LOG: ${log_entry}
-                echo POD: ${pod}
+            for log_entry in "${pattern_1}" "${pattern_2}"; do
+                if [[ -z ${log_entry} ]]; then
+                    echo "No Conflicting LogEntry on ${conflicting_daemon_pod}"
+                else
+                    echo ">> Conflicting LogEntry Found!!"
+                    pod=$(echo ${log_entry##*pods/} | cut -d\" -f2)
+                    conflicting_ns=$(oc --kubeconfig=${KUBEC} get pod -A | grep ${pod} | cut -f1 -d\ )
 
-                oc --kubeconfig=${KUBEC} delete pod -n ${conflicting_ns} ${pod} --force --grace-period=0
-            fi
-        done
+                    echo ">> Clean Eviction triggered info: "
+                    echo NODE: ${conflicting_node}
+                    echo DAEMON: ${conflicting_daemon_pod}
+                    echo NS: ${conflicting_ns}
+                    echo LOG: ${log_entry}
+                    echo POD: ${pod}
 
+                    oc --kubeconfig=${KUBEC} delete pod -n ${conflicting_ns} ${pod} --force --grace-period=0
+                fi
+            done
+        fi
     fi
 }
